seperate socket in seperate file
