+++ server/index.js
@@ -1,4 +1,5 @@
-const { server, app } = require('./app');
+const app = require('./app');
+const server = require('./socket');
 
 server.listen(app.get('PORT'), () => {
   console.log('http://localhost:5000');
